move from_iter_insteam_of_collect to its own module
